Update 1.1.0
Fixed duplicated . in the ReadMe file.
Included the missing [INFO] tag in the Unified Audit Log extraction script.
Introduced introductory text in the Read The Docs documentation for transport rules.
Enhanced JSON functionality by adding JSON depth to the ConvertTo-Json feature in the Get-AzureLogs script.
Implemented a retry counter in the Unified Audit Log extraction script.
Renamed the Get-AzureLogs.ps1 script to Get-AzureADLogs.ps1 for better clarity.
(BETA) Introduced the Get-AzureActivityLogs script to enable acquisition of Azure Activity logging.
Modified certain commands in the Get-AzureADLogs.ps1 script for improved functionality.
Added the Azure Activity Log section to the Read The Docs documentation.
Expanded the supported sources in Read The Docs to include Activity logs.
Updated the installation step in Read The Docs to include the AZ module.
(BETA) Integrated a script for acquiring Sign-in logs using the Graph API.
(BETA) Integrated a script for acquiring Audit logs using the Graph API.
Fixed column formatting in JSON output for the Get-ADAuditLogs command.
Included the Graph API in the connect script.
Added a note about the StartDate and EndDate parameters for enhanced clarity.